Main Loop for Gesture Recognition
This part comprises the main loop of the program. It continuously captures frames from the webcam, processes them for hand detection, and recognizes gestures based on the number of fingers raised.

If a hand is detected, it calculates the number of fingers raised using the ' count_fingers() ' function.

It then compares the current finger count with the previous count to avoid repeated actions and ensures a minimum time interval between consecutive actions.

Depending on the finger count, it triggers specific keyboard commands through ' pyautogui '

It draws landmarks and connections on the captured frame to visualize hand detection using  ' MediaPipe '

The loop exits when the 'Esc' key is pressed, releasing resources and closing windows.